moved roimaskselector to util
