+++ snap-rcp/src/main/java/org/esa/snap/rcp/statistics/ScatterPlot3DPlotPanel.java
@@ -21,7 +21,6 @@ import com.bc.ceres.binding.PropertyContainer;
 import com.bc.ceres.binding.ValidationException;
 import com.bc.ceres.binding.ValueRange;
 import com.bc.ceres.binding.ValueSet;
-import com.bc.ceres.core.ProgressMonitor;
 import com.bc.ceres.glevel.MultiLevelImage;
 import com.bc.ceres.glevel.MultiLevelModel;
 import com.bc.ceres.glevel.support.AbstractMultiLevelSource;
@@ -65,13 +64,13 @@ import org.esa.snap.core.datamodel.ImageInfo;
 import org.esa.snap.core.datamodel.Mask;
 import org.esa.snap.core.datamodel.Product;
 import org.esa.snap.core.datamodel.RasterDataNode;
-import org.esa.snap.core.datamodel.Stx;
 import org.esa.snap.core.image.FillConstantOpImage;
 import org.esa.snap.core.image.ImageManager;
 import org.esa.snap.core.util.Debug;
 import org.esa.snap.rcp.SnapApp;
 import org.esa.snap.rcp.colormanip.ScatterPlot3DColorManipulationPanel;
 import org.esa.snap.rcp.colormanip.ScatterPlot3DFormModel;
+import org.esa.snap.rcp.util.RoiMaskSelector;
 import org.esa.snap.ui.AbstractDialog;
 import org.esa.snap.ui.GridBagUtils;
 import org.esa.snap.ui.UIUtils;
